#50: Refactor to rd_credentials.py.
