+++ src/raindrop_todoist_syncer/raindrop.py
@@ -1,4 +1,3 @@
-import json
 import os
 from loguru import logger
 import shutil
@@ -7,7 +6,7 @@ from typing import Any, Dict, List
 
 from dotenv import load_dotenv
 import requests
-from requests import Request, Response
+from requests import Response
 from tenacity import (
     retry,
     retry_if_exception_type,
@@ -509,51 +508,3 @@ class EnvironmentVariablesFileManager:
         shutil.copy(self.env_file, self.env_backup)
         with open(self.env_file, "w") as file:
             file.writelines(validated_new_body)
-
-
-class RaindropCredentialsManager:
-    def __init__(self) -> None:
-        """Provide variables and methods for managing Raindrop Oauth2 credentials."""
-        self.env_file = ".env"
-        self.AUTH_CODE_BASE_URL = "https://raindrop.io/oauth/authorize"
-        self.REDIRECT_URI = "http://localhost"
-        self.HEADERS = {"Content-Type": "application/json"}
-        self.RAINDROP_CLIENT_ID = os.getenv("RAINDROP_CLIENT_ID")
-        self.RAINDROP_CLIENT_SECRET = os.getenv("RAINDROP_CLIENT_SECRET")
-        self.RAINDROP_REFRESH_TOKEN = os.getenv("RAINDROP_REFRESH_TOKEN")
-        self.RAINDROP_ACCESS_TOKEN = os.getenv("RAINDROP_ACCESS_TOKEN")
-
-    def make_request(self, body: Dict[str, str]) -> Request:
-        """Makes the an request and returns a Request object."""
-        headers = self.HEADERS
-        data = body
-        oauth_response = requests.post(
-            "https://raindrop.io/oauth/access_token",
-            headers=headers,
-            data=json.dumps(data),
-        )
-        return oauth_response
-
-    def response_validator(self, response: Response) -> None:
-        """Checks a Response object returned by the Raindrop API Oauth2 process is
-        valid.
-        """
-        if response.status_code != 200:
-            raise ValueError(
-                "Response status code is not 200 (as required in the docs)."
-                f"Status code was {response.status_code} - {response.text}"
-            )
-
-        if response.json().get("access_token") is None:
-            raise ValueError(
-                f"Response code 200 but no token in response. Full response {response.json()}"
-            )
-
-    def extract_access_token(self, oauth_response: Response) -> str:
-        """Extracts the access token from the response.json of a Response object."""
-        data = oauth_response.json()
-        access_token = data.get("access_token")
-        logger.info(
-            f"Your access token is {access_token}. I am of type {type(access_token)}"
-        )
-        return access_token
