+++ tests/conftest.py
@@ -7,10 +7,8 @@ import shutil
 import pytest
 from requests import HTTPError
 
-from raindrop_todoist_syncer.raindrop import (
-    RaindropCredentialsManager,
-    EnvironmentVariablesFileManager,
-)
+from raindrop_todoist_syncer.raindrop import EnvironmentVariablesFileManager
+from raindrop_todoist_syncer.rd_credentials import RaindropCredentialsManager
 from raindrop_todoist_syncer.rd_token import RaindropAccessTokenRefresher
 from raindrop_todoist_syncer.rd_class import Raindrop
