+++ tests/unit/test_raindrop_credentials_manager.py
@@ -2,7 +2,7 @@ from unittest.mock import patch
 
 import pytest
 
-from raindrop_todoist_syncer.raindrop import RaindropCredentialsManager
+from raindrop_todoist_syncer.rd_credentials import RaindropCredentialsManager
 
 
 @pytest.fixture
